damage-ai-agent/requirements.txt
streamlit
ultralytics
opencv-python
numpy
pillow
requests
